+++ modules/apps/bookmarks/bookmarks-uad-test/src/testIntegration/java/com/liferay/bookmarks/uad/test/util/BookmarksEntryUADTestUtil.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.bookmarks.uad.test;
+package com.liferay.bookmarks.uad.test.util;
 
 import com.liferay.bookmarks.constants.BookmarksFolderConstants;
 import com.liferay.bookmarks.model.BookmarksEntry;
